Implement an interface for validation service
