+++ src/test/java/io/github/christophermanahan/captainlunch/web/UserControllerTest.java
@@ -2,7 +2,7 @@ package io.github.christophermanahan.captainlunch.web;
 
 import io.github.christophermanahan.captainlunch.model.User;
 import io.github.christophermanahan.captainlunch.service.CreateUserService;
-import io.github.christophermanahan.captainlunch.service.ValidationService;
+import io.github.christophermanahan.captainlunch.service.SlackSigningSecretValidationService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -29,14 +29,14 @@ class UserControllerTest {
     private CreateUserService userService;
 
     @MockBean
-    private ValidationService validationService;
+    private SlackSigningSecretValidationService slackSigningSecretValidationService;
 
     @Test
     @SuppressWarnings("unchecked")
     void createsUserIfRequestIsValid() throws Exception {
         String identity = "W100000";
         when(userService.createUser(identity)).thenReturn(new User(identity, new Date(Long.valueOf("0"))));
-        when(validationService.validateRequest(any(HttpEntity.class))).thenReturn(true);
+        when(slackSigningSecretValidationService.validateRequest(any(HttpEntity.class))).thenReturn(true);
 
         mvc.perform(post("/join")
             .contentType(MediaType.APPLICATION_FORM_URLENCODED)
@@ -50,7 +50,7 @@ class UserControllerTest {
     void failsToCreatesUserIfRequestIsInvalid() throws Exception {
         String identity = "W100000";
         when(userService.createUser(identity)).thenReturn(new User(identity, new Date(Long.valueOf("0"))));
-        when(validationService.validateRequest(any(HttpEntity.class))).thenReturn(false);
+        when(slackSigningSecretValidationService.validateRequest(any(HttpEntity.class))).thenReturn(false);
 
         mvc.perform(post("/join")
                 .contentType(MediaType.APPLICATION_FORM_URLENCODED)
@@ -64,7 +64,7 @@ class UserControllerTest {
     void isFoundIfUserAlreadyExist() throws Exception {
         String identity = "W100000";
         when(userService.createUser(identity)).thenThrow(DataIntegrityViolationException.class);
-        when(validationService.validateRequest(any(HttpEntity.class))).thenReturn(true);
+        when(slackSigningSecretValidationService.validateRequest(any(HttpEntity.class))).thenReturn(true);
 
         mvc.perform(post("/join")
                 .contentType(MediaType.APPLICATION_FORM_URLENCODED)
